+++ application/webserver/endpoints/handler_about.go
@@ -2,7 +2,7 @@ package gatesentryWebserverEndpoints
 
 import gatesentryWebserverTypes "bitbucket.org/abdullah_irfan/gatesentryf/webserver/types"
 
-func GSApiAboutGET(runtime *gatesentryWebserverTypes.TemporaryRuntime) interface{} {
+func GSApiAboutGET(runtime *gatesentryWebserverTypes.TemporaryRuntime, adminHTTPSPort string) interface{} {
 	apikey := runtime.GetInstallationId()
 	usagedata, msg := runtime.GetTotalConsumptionData()
 	version := runtime.GetApplicationVersion()
@@ -12,5 +12,6 @@ func GSApiAboutGET(runtime *gatesentryWebserverTypes.TemporaryRuntime) interface
 		Usagedata           string
 		Additionalusagedata string
 		Version             string `json:"version"`
-	}{Apikey: apikey, Usagedata: usagedata, Additionalusagedata: msg, Version: version}
+		AdminHTTPSPort      string `json:"admin_https_port"`
+	}{Apikey: apikey, Usagedata: usagedata, Additionalusagedata: msg, Version: version, AdminHTTPSPort: adminHTTPSPort}
 }